Restructure and move response classes to own file
